Refactoring - chat -> forum
